pyproject.toml
[tool.poetry]
name = "realestate-multimodal-ai"
version = "0.1.0"
description = ""
authors = ["Boyd Sorratat <redthegx@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
transformers = "^4.46.3"
timm = "^1.0.11"
faiss-cpu = "^1.9.0.post1"
numpy = "^2.1.3"
pandas = "^2.2.3"
scikit-learn = "^1.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
